+++ test/fixes/convert_to_ternary_fix_test.dart
@@ -1,5 +1,5 @@
-import 'package:custom_lints/src/fixes/convert_to_ternary_fix.dart';
-import 'package:custom_lints/src/rules/prefer_ternary.dart';
+import 'package:code_craft_lints/src/fixes/convert_to_ternary_fix.dart';
+import 'package:code_craft_lints/src/rules/prefer_ternary.dart';
 import 'package:test_reflective_loader/test_reflective_loader.dart';
 
 import 'fix_test_.dart';