+++ test/rules/prefer_ternary_test.dart
@@ -1,5 +1,5 @@
 import 'package:analyzer_testing/analysis_rule/analysis_rule.dart';
-import 'package:custom_lints/src/rules/prefer_ternary.dart';
+import 'package:code_craft_lints/src/rules/prefer_ternary.dart';
 import 'package:test_reflective_loader/test_reflective_loader.dart';
 
 void main() {